DT_Kiwi/requirements.txt
# Core test framework dependencies
playwright
pytest
pytest-playwright
pytest-xdist

# Test ordering
pytest-order

# Kiwi TCMS integration
tcms-api
kiwitcms

# Screen resolution management
screeninfo

# Optional utilities
python-dotenv
loguru
requests
